requirements.txt
numpy==1.24.3
scipy
tensorflow==2.9.1
tensorflow-probability==0.17.0
pandas
python-dotenv
scikit-learn